# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Home page.",
                "summary": "Home Page",
                "responses": {
                    "200": {
                        "description": "home page",
                        "schema": {
                            "type": "html"
                        }
                    }
                }
            }
        },
        "/_": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Admin route.",
                "summary": "Admin route",
                "responses": {
                    "200": {
                        "description": "Admin page",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get cart.",
                "summary": "Get cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetCartResponse"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carts/{productId}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add to cart.",
                "summary": "Add to cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "400": {
                        "description": "invalid product",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete cart item.",
                "summary": "Delete cart item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "400": {
                        "description": "invalid product",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carts/{productId}/{quantity}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update cart item quantity.",
                "summary": "Update cart item quantity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "productId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "400": {
                        "description": "invalid product or quantity",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/config": {
            "get": {
                "description": "Get client config.",
                "summary": "Get config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/coupons": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get available coupons.",
                "summary": "Get available coupons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetAvailableCouponsResponse"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/coupons/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all coupons.",
                "summary": "Get all coupons",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetAllCouponsResponse"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user.",
                "summary": "Get user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.User"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create order.",
                "summary": "Create order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coupon code",
                        "name": "couponCode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateOrderResponse"
                        }
                    },
                    "400": {
                        "description": "invalid coupon",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all orders.",
                "summary": "Get all orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetAllOrdersResponse"
                        }
                    },
                    "401": {
                        "description": "invalid session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get products.",
                "summary": "Get products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetProductsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateOrderResponse": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/repo.Order"
                }
            }
        },
        "handler.GetAllCouponsResponse": {
            "type": "object",
            "properties": {
                "coupons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.Coupon"
                    }
                }
            }
        },
        "handler.GetAllOrdersResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.Order"
                    }
                }
            }
        },
        "handler.GetAvailableCouponsResponse": {
            "type": "object",
            "properties": {
                "coupons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.Coupon"
                    }
                }
            }
        },
        "handler.GetCartResponse": {
            "type": "object",
            "properties": {
                "cart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.CartItem"
                    }
                }
            }
        },
        "handler.GetProductsResponse": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.Product"
                    }
                }
            }
        },
        "handler.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "repo.CartItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "repo.Coupon": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "discountPercent": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isUsed": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "repo.Order": {
            "type": "object",
            "properties": {
                "couponId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "discountedAmount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "repo.Product": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "description": "Price is in the smallest unit of the currency",
                    "type": "integer"
                },
                "quantityLeft": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "repo.User": {
            "type": "object",
            "properties": {
                "accountStatus": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}